+++ examples/bitmap_palette.rs
@@ -9,7 +9,8 @@ fn main() -> ! {
     let mut bus = unsafe { Mmio::new() };
     let mut bitmap = Bitmap::new(&mut bus);
     for y in 0..32 { for x in 0..32 {
-        bitmap.pixel(x,y,Color::from_bits(((x / 4) as u8) << 5));
+        // 16 by 16 color swatches, each two pixels wide and high.
+        bitmap.pixel(x,y,Color::from_bits(((y / 2) * 16 + x / 2) as u8));
     }}
     os::runtime::halt()
 }